backend/requirements.txt
Flask==2.3.2
Flask-CORS==4.0.0
mysql-connector-python==8.0.33
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.2
urllib3==2.0.4
